pyproject.toml
[project]
name = "ad-insights"
version = "0.1.0"
description = "Ad campaign insights generator"
requires-python = ">=3.11"
dependencies = [
    "polars>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "fastexcel>=0.11.0",  # Excel read support for polars
    "scipy>=1.11.0",      # Statistical functions
    "numpy>=1.24.0",      # Array operations for scipy
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-cov>=4.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
